KAFKA-12553: Refactor recovery logic
