+++ core/src/test/scala/unit/kafka/cluster/PartitionLockTest.scala
@@ -344,14 +344,17 @@ class PartitionLockTest extends Logging {
   private class SlowLog(log: Log, mockTime: MockTime, appendSemaphore: Semaphore) extends Log(
     log.dir,
     log.config,
+    log.segments,
     log.logStartOffset,
     log.recoveryPoint,
+    log.nextOffsetMetadata,
     mockTime.scheduler,
     new BrokerTopicStats,
     log.time,
     log.maxProducerIdExpirationMs,
     log.producerIdExpirationCheckIntervalMs,
     log.topicPartition,
+    log.leaderEpochCache,
     log.producerStateManager,
     new LogDirFailureChannel(1),
     topicId = None,
