+++ core/src/test/scala/unit/kafka/cluster/PartitionTest.scala
@@ -1937,14 +1937,17 @@ class PartitionTest extends AbstractPartitionTest {
   private class SlowLog(log: Log, mockTime: MockTime, appendSemaphore: Semaphore) extends Log(
     log.dir,
     log.config,
+    log.segments,
     log.logStartOffset,
     log.recoveryPoint,
+    log.nextOffsetMetadata,
     mockTime.scheduler,
     new BrokerTopicStats,
     log.time,
     log.maxProducerIdExpirationMs,
     log.producerIdExpirationCheckIntervalMs,
     log.topicPartition,
+    log.leaderEpochCache,
     log.producerStateManager,
     new LogDirFailureChannel(1),
     topicId = None,
