+++ core/src/test/scala/unit/kafka/utils/SchedulerTest.scala
@@ -19,7 +19,8 @@ package kafka.utils
 import java.util.Properties
 import java.util.concurrent.atomic._
 import java.util.concurrent.{CountDownLatch, Executors, TimeUnit}
-import kafka.log.{Log, LogConfig, LogManager, ProducerStateManager}
+
+import kafka.log.{Log, LogConfig, LogLoader, LogManager}
 import kafka.server.{BrokerTopicStats, LogDirFailureChannel}
 import kafka.utils.TestUtils.retry
 import org.junit.jupiter.api.Assertions._
@@ -118,13 +119,16 @@ class SchedulerTest {
     val logDir = TestUtils.randomPartitionLogDir(tmpDir)
     val logConfig = LogConfig(new Properties())
     val brokerTopicStats = new BrokerTopicStats
-    val recoveryPoint = 0L
     val maxProducerIdExpirationMs = 60 * 60 * 1000
     val topicPartition = Log.parseTopicPartitionName(logDir)
-    val producerStateManager = new ProducerStateManager(topicPartition, logDir, maxProducerIdExpirationMs)
-    val log = new Log(logDir, logConfig, logStartOffset = 0, recoveryPoint = recoveryPoint, scheduler,
+    val logDirFailureChannel = new LogDirFailureChannel(10)
+    val logLoader = new LogLoader(logDir, topicPartition, logConfig, scheduler, mockTime, logDirFailureChannel)
+    val logComponents = logLoader.load(0L, 0L, maxProducerIdExpirationMs)
+    val log = new Log(logDir, logConfig, segments = logComponents.segments, logStartOffset = logComponents.logStartOffset,
+      recoveryPoint = logComponents.recoveryPoint, nextOffsetMetadata = logComponents.nextOffsetMetadata, scheduler,
       brokerTopicStats, mockTime, maxProducerIdExpirationMs, LogManager.ProducerIdExpirationCheckIntervalMs,
-      topicPartition, producerStateManager, new LogDirFailureChannel(10), topicId = None, keepPartitionMetadataFile = true)
+      topicPartition, logComponents.leaderEpochCache, logComponents.producerStateManager, logDirFailureChannel,
+      topicId = None, keepPartitionMetadataFile = true)
     assertTrue(scheduler.taskRunning(log.producerExpireCheck))
     log.close()
     assertFalse(scheduler.taskRunning(log.producerExpireCheck))
